added security

+++ src/main/java/com/nstanogias/knapsack/service/KnapsackService.java
@@ -1,13 +1,11 @@
 package com.nstanogias.knapsack.service;
 
-
 import com.nstanogias.knapsack.model.Knapsack;
-
 import java.util.Optional;
 
 public interface KnapsackService {
 
-    Optional<Knapsack> getKnapsack(Long task);
+    Optional<Knapsack> getKnapsack(Integer task);
 
     Knapsack storeKnapsack(Knapsack knapsack);
